+++ redwood-layout-uiview/src/commonTest/kotlin/app/cash/redwood/layout/uiview/UIViewFlexContainerTest.kt
@@ -22,8 +22,10 @@ import app.cash.redwood.layout.widget.FlexContainer
 import app.cash.redwood.widget.ChangeListener
 import app.cash.redwood.widget.Widget
 import app.cash.redwood.yoga.FlexDirection
+import kotlinx.cinterop.cValue
 import platform.CoreGraphics.CGRectMake
 import platform.UIKit.UIColor
+import platform.UIKit.UIScrollView
 import platform.UIKit.UIView
 
 class UIViewFlexContainerTest(
@@ -50,10 +52,19 @@ class UIViewFlexContainerTest(
     FlexContainer<UIView> by delegate,
     ChangeListener by delegate {
     private var childCount = 0
+
     override val children: Widget.Children<UIView> = delegate.children
 
     init {
       value.backgroundColor = UIColor(red = 0.0, green = 0.0, blue = 1.0, alpha = 0.2)
+    }
+
+    override fun onScroll(onScroll: ((Double) -> Unit)?) {
+      delegate.onScroll(onScroll)
+    }
+
+    override fun scroll(offset: Double) {
+      (delegate.value as UIScrollView).setContentOffset(cValue { y = offset }, false)
     }
 
     override fun add(widget: Widget<UIView>) {